fix(reference): fix the reference-summary panic on a completed import
GET /api/v1/reference/summary (the counts the Settings tiles show) crashed
the request whenever any kind had a row in reference_imports: it read
rows_added/rows_updated/rows_deactivated as i64, but those columns are
INTEGER, not BIGINT. sqlx's typed get() panics rather than coerces on a
width mismatch, which took down the connection (curl saw "empty reply from
server", not even a 500).

This is why it only showed up now: reference_imports could not gain a row
before the init.sql fix for the missing reference-code tables, since apply
mode's own INSERT into that table needs icd10_codes/cpt_codes/etc. to
exist first. First real import, first row in reference_imports, first hit
of this.

+++ crates/api-gateway/src/routes/reference.rs
@@ -703,9 +703,13 @@ pub async fn reference_summary(State(state): State<AppState>) -> Result<Json<Val
                 let at: chrono::DateTime<Utc> = l.get("imported_at");
                 let by: Option<String> = l.get("imported_by");
                 let filename: Option<String> = l.get("filename");
-                let added: i64 = l.get("rows_added");
-                let updated: i64 = l.get("rows_updated");
-                let deactivated: i64 = l.get("rows_deactivated");
+                // These columns are INTEGER, not BIGINT - sqlx's typed get()
+                // panics rather than coerces on a width mismatch, which this
+                // path never surfaced before nothing could reach it (see the
+                // init.sql fix for the reference-code tables it reads from).
+                let added: i32 = l.get("rows_added");
+                let updated: i32 = l.get("rows_updated");
+                let deactivated: i32 = l.get("rows_deactivated");
                 json!({
                     "at": at.to_rfc3339(),
                     "by": by,
